refactor(1013): remove DeviceRebootManager facade (SC-001, position 41)
Extract DeviceRebootManager (~397 LOC) from MistHelper.py to
src/device/device_reboot_manager.py using the SC-001 facade pattern.
Add re-export alias in MistHelper.py's src.device import block and
replace the class body with a single-line breadcrumb comment. Update
pyproject.toml coverage omit list with the new module.

Local gate: ruff clean, black --check clean, py_compile OK,
python MistHelper.py --test = 66 pass / 0 fail / 131 skip (baseline).
